requirements.txt
django
djangorestframework
djangorestframework-simplejwt
numpy
scikit-learn
reportlab